upd(table): use common logic
